Delete dead files
We are not going to work on Rhapsody, and, this being Objective-C,
the code is unlikely to help us in any way. So, we remove the
entire directory.

Also, there are `makezip' and `makezip.bat' files that apparently
were created by the previous developers as a convenience tool.
We are not using zip, so we let the file go.