Remove unused `qport' local variable
Found by GCC 5.4.0: a local `qport' variable, which by
the way shadows a global variable with the same name,
is set but not used. Although the intention behind the
offending variable is not knows, we believe that the
local variable can safely be removed, and we do remove
it in this patch.